Titanic2/Backend-Titanic/requirements.txt
pandas
numpy
joblib
scikit-learn
Flask # web server
tensorflow-cpu
gunicorn 
